+++ include/character_listings.h
@@ -25,6 +25,27 @@
 #define STAIRCASE_UP_CHARACTER '<'
 #define STAIRCASE_DOWN_CHARACTER '>'
 
+// Object characters
+#define WEAPON_CHARACTER        '|'
+#define OFFHAND_CHARACTER       ')'
+#define RANGED_CHARACTER        '}'
+#define ARMOR_CHARACTER         '['
+#define HELMET_CHARACTER        ']'
+#define CLOAK_CHARACTER         '('
+#define GLOVES_CHARACTER        '{'
+#define BOOTS_CHARACTER         '\\'
+#define RING_CHARACTER          '='
+#define AMULET_CHARACTER        '"'
+#define LIGHT_CHARACTER         '_'
+#define SCROLL_CHARACTER        '~'
+#define BOOK_CHARACTER          '?'
+#define FLASK_CHARACTER         '!'
+#define GOLD_CHARACTER          '$'
+#define AMMUNITION_CHARACTER    '/'
+#define FOOD_CHARACTER          ','
+#define WAND_CHARACTER          '-'
+#define CONTAINER_CHARACTER     '%'
+
 #define ROCK_HARDNESS_MIN 50
 #define ROCK_HARDNESS_MAX 225
 #define BORDER_HARDNESS 255